requirements.txt
textual>=0.85
openai>=1.50
google-genai>=0.7
python-dotenv>=1.0
